refactor:move some code from pkg/event to pkg/event/schemas
